Move resolveNodeModuleDir to cli-tools package
