+++ packages/cli/src/tools/config/__tests__/findDependencies-test.ts
@@ -5,8 +5,6 @@ import {
   getTempDirectory,
 } from '../../../../../../jest/helpers';
 
-jest.mock('../resolveNodeModuleDir');
-
 beforeEach(async () => {
   await cleanup(DIR);
   jest.resetModules();
